+++ libs/rxjs/boolean/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `fromBoolean` no longer accept arguments list or Promise values
+
 ## [3.2.1] - 2021-01-18
 
 ### Changed